requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=0.20.0
SQLAlchemy>=2.0.0
httpx>=0.24.0
boto3>=1.26.0
openai>=1.0.0
elevenlabs>=0.2.0
requests>=2.28.0
python-multipart>=0.0.6